chore: convert HeaderNavigation to TS
