+++ src/config/constants.ts
@@ -1,4 +1,9 @@
-import { buildSignInPath, Context, ItemType, buildPdfViewerLink } from '@graasp/sdk';
+import {
+  Context,
+  ItemType,
+  buildPdfViewerLink,
+  buildSignInPath,
+} from '@graasp/sdk';
 
 export const APP_NAME = 'Graasp';
 
@@ -9,9 +14,7 @@ export const ENV = {
 };
 
 export const NODE_ENV =
-  process.env.REACT_APP_NODE_ENV ||
-  process.env.NODE_ENV ||
-  ENV.DEVELOPMENT;
+  process.env.REACT_APP_NODE_ENV || process.env.NODE_ENV || ENV.DEVELOPMENT;
 
 export const SENTRY_DSN = process.env.REACT_APP_SENTRY_DSN;
 
@@ -21,30 +24,25 @@ export const API_HOST =
 export const DOMAIN = process.env.REACT_APP_DOMAIN;
 
 export const SHOW_NOTIFICATIONS =
-  process.env.REACT_APP_SHOW_NOTIFICATIONS === 'true' ||
-  false;
+  process.env.REACT_APP_SHOW_NOTIFICATIONS === 'true' || false;
 
 export const AUTHENTICATION_HOST =
-  process.env.REACT_APP_AUTHENTICATION_HOST ||
-  'http://localhost:3001';
+  process.env.REACT_APP_AUTHENTICATION_HOST || 'http://localhost:3001';
 
 export const GRAASP_COMPOSE_HOST =
-  process.env.REACT_APP_GRAASP_COMPOSE_HOST ||
-  'http://localhost:3111';
+  process.env.REACT_APP_GRAASP_COMPOSE_HOST || 'http://localhost:3111';
 
 export const GRAASP_EXPLORE_HOST =
-  process.env.REACT_APP_GRAASP_EXPLORE_HOST ||
-  'http://localhost:3005';
+  process.env.REACT_APP_GRAASP_EXPLORE_HOST || 'http://localhost:3005';
 
 export const GRAASP_ANALYTICS_HOST =
-  process.env.REACT_APP_GRAASP_ANALYTICS_HOST ||
-  'http://localhost:3012';
+  process.env.REACT_APP_GRAASP_ANALYTICS_HOST || 'http://localhost:3012';
 
 export const H5P_INTEGRATION_URL =
-  process.env.REACT_APP_H5P_INTEGRATION_URL ||
-  `${API_HOST}/p/h5p-integration`;
+  process.env.REACT_APP_H5P_INTEGRATION_URL || `${API_HOST}/p/h5p-integration`;
 
-export const HIDDEN_ITEM_TAG_ID = process.env.REACT_APP_HIDDEN_ITEM_TAG_ID || false;
+export const HIDDEN_ITEM_TAG_ID =
+  process.env.REACT_APP_HIDDEN_ITEM_TAG_ID || false;
 
 export const GA_MEASUREMENT_ID = process.env.REACT_APP_GA_MEASUREMENT_ID;
 
@@ -84,7 +82,9 @@ export const buildBuilderTabName = (id: string): string => `builder-tab-${id}`;
 export const DEFAULT_LINK_SHOW_BUTTON = true;
 export const DEFAULT_RESIZABLE_SETTING = false;
 
-export const PDF_VIEWER_LINK = buildPdfViewerLink(process.env.REACT_APP_GRAASP_ASSETS_URL)
+export const PDF_VIEWER_LINK = buildPdfViewerLink(
+  process.env.REACT_APP_GRAASP_ASSETS_URL,
+);
 
 // todo: remove when files that use this are in TS
-export const APP_CONTEXT = Context.PLAYER
+export const APP_CONTEXT = Context.PLAYER;
